refactor routes (guest, auth)
